requirements.txt
numpy
tqdm
six
tensorflow==1.13.1
tensorflow-gpu==1.13.1
tensorflow_probability==0.6.0
pypianoroll==0.5.0
PyYAML